fix: fix tests to API
